+++ backend/sample.env
@@ -65,7 +65,7 @@ API_STORAGE_DIR = "/data/api"
 PLATFORM_SERVICE_HOST=http://unstract-platform-service
 PLATFORM_SERVICE_PORT=3001
 
-# Tool Runner (Worker Service)
+# Tool Runner
 UNSTRACT_RUNNER_HOST=http://unstract-runner
 UNSTRACT_RUNNER_PORT=5002
