+++ docker/dockerfiles/frontend.Dockerfile.dockerignore
@@ -59,5 +59,5 @@ platform-service
 prompt-service
 tools
 unstract
-worker
+runner
 x2text-service
