+++ docker/dockerfiles/platform.Dockerfile.dockerignore
@@ -59,5 +59,5 @@ frontend
 prompt-service
 tools
 !unstract
-worker
+runner
 x2text-service
