+++ docker/dockerfiles/runner.Dockerfile
@@ -8,7 +8,7 @@ ENV PYTHONDONTWRITEBYTECODE 1
 ENV PYTHONUNBUFFERED 1
 ENV PYTHONPATH /unstract
 
-ENV BUILD_CONTEXT_PATH worker
+ENV BUILD_CONTEXT_PATH runner
 ENV BUILD_PACKAGES_PATH unstract
 ENV PDM_VERSION 2.16.1
