+++ runner/sample.env
@@ -9,7 +9,7 @@ PRIVATE_REGISTRY_CREDENTIAL_PATH=
 PRIVATE_REGISTRY_USERNAME=
 PRIVATE_REGISTRY_URL=
 
-# Log level for worker (Options: INFO, WARNING, ERROR, DEBUG, etc.)
+# Log level for runner (Options: INFO, WARNING, ERROR, DEBUG, etc.)
 LOG_LEVEL="INFO"
 
 # Flag to decide whether to clean up/ remove the tool container after execution.
@@ -17,7 +17,7 @@ LOG_LEVEL="INFO"
 REMOVE_CONTAINER_ON_EXIT=True
 
 # Client module path of the container engine to be used.
-CONTAINER_CLIENT_PATH=unstract.worker.clients.docker
+CONTAINER_CLIENT_PATH=unstract.runner.clients.docker
 
 EXECUTION_RUN_DATA_FOLDER_PREFIX="/app/workflow_data"
