+++ runner/src/unstract/runner/main.py
@@ -1,8 +1,8 @@
 from typing import Any, Optional
 
 from flask import Blueprint, Flask, Response, abort, jsonify, request
-from unstract.worker import UnstractWorker
-from unstract.worker.utils import Utils
+from unstract.runner import UnstractRunner
+from unstract.runner.utils import Utils
 
 app = Flask(__name__)
 
@@ -33,8 +33,8 @@ def run_container() -> Optional[Any]:
     envs = data["envs"]
     messaging_channel = data["messaging_channel"]
 
-    worker = UnstractWorker(image_name, image_tag, app)
-    result = worker.run_container(
+    runner = UnstractRunner(image_name, image_tag, app)
+    result = runner.run_container(
         organization_id=organization_id,
         workflow_id=workflow_id,
         execution_id=execution_id,
@@ -61,9 +61,9 @@ def run_command(command: str) -> Optional[Any]:
         abort(404)
     image_name = request.args.get("image_name")
     image_tag = request.args.get("image_tag")
-    worker = UnstractWorker(image_name, image_tag, app)
+    runner = UnstractRunner(image_name, image_tag, app)
 
-    return worker.run_command(command)
+    return runner.run_command(command)
 
 
 # Register the Blueprint with the Flask app
