+++ tox.ini
@@ -1,11 +1,11 @@
 [tox]
-env_list = py{39,310,311}, worker
+env_list = py{39,310,311}, runner
 
 # [testenv]
 # skip_install = true
 
-[testenv:worker]
-changedir = worker
+[testenv:runner]
+changedir = runner
 setenv =
     PDM_IGNORE_SAVED_PYTHON="1"
     PROTOCOL_BUFFERS_PYTHON_IMPLEMENTATION=python
@@ -16,4 +16,4 @@ commands_pre =
     pdm sync --dev
     sh -c '[ -f cloud_requirements.txt ] && pip install -r cloud_requirements.txt || echo "cloud_requirements.txt not found"'
 commands =
-    pytest -v --md-report-verbose=1 --md-report --md-report-flavor gfm --md-report-output ../worker-report.md
+    pytest -v --md-report-verbose=1 --md-report --md-report-flavor gfm --md-report-output ../runner-report.md
