--- worker/src/unstract/worker/__init__.py
@@ -1,3 +1,0 @@
-from .worker import UnstractWorker
-
-__all__ = ["UnstractWorker"]
